requirements.txt
numpy==1.26.0
pandas==2.2.2
scikit-learn==1.5.1
tensorflow==2.17.0
pydantic==2.8.2
streamlit==1.37.1
scipy==1.14.1
pyyaml==6.0.2
matplotlib==3.9.2
xgboost==2.1.1
fastapi==0.112.2
uvicorn==0.30.6
langchain-openai
markdown2
python-multipart
seaborn
statsmodels
causalinference
dowhy
featuretools
econml